第14部分-复习/day152-下午村长分享/otherDemo/utils/form.py
# -*- coding: utf-8 -*-
# __author__ = "maple"

from django import forms

from ckeditor.widgets import CKEditorWidget

from web import models


class CommentForm(forms.ModelForm):
    content = forms.CharField(widget=CKEditorWidget())

    class Meta:
        fields = "__all__"
        model = models.Comment
