第14部分-复习/day152-下午村长分享/otherDemo/requirements.txt
django==1.11.3
django-ckeditor==5.5.0
pymysql==0.7.11
sentry-sdk==0.7.3
requests==2.18.1